fix(node:fs): pass ENAMETOOLONG to callback instead of throwing
When fs.readFile() is called with a path that exceeds the OS limit
(typically 1024 chars), the error was thrown synchronously instead
of being passed to the callback asynchronously.

This fix wraps the internal promise call in a try-catch and uses
process.nextTick to pass any synchronous errors to the callback.

Fixes #25659

Signed-off-by: majiayu000 <ma_jiayu@outlook.com>
Signed-off-by: lif <1835304752@qq.com>