BuildBun.cmake: simplify bindgen invocation using COMMAND_ERROR_IS_FATAL
In CMake, failure to execute a process will place a message string in the RESULT_VARIABLE. If the message string starts with 'No' such as in 'No such file or directory' then CMake will interpret that as the boolean false and not halt the build. The new code uses a built-in option to halt the build on any failure, so the script will halt correctly if that error occurs. This could also be fixed by quoting.